MYSHOP-46 Added permission before performing operations

+++ app/src/main/java/com/labs/devo/apps/myshop/business/notebook/implementation/EntryRepositoryImpl.kt
@@ -1,6 +1,9 @@
 package com.labs.devo.apps.myshop.business.notebook.implementation
 
+import com.labs.devo.apps.myshop.business.helper.PermissionsHelper
+import com.labs.devo.apps.myshop.business.helper.PermissionsHelper.checkPermissions
 import com.labs.devo.apps.myshop.business.notebook.abstraction.EntryRepository
+import com.labs.devo.apps.myshop.const.Permissions
 import com.labs.devo.apps.myshop.data.db.local.abstraction.notebook.LocalEntryService
 import com.labs.devo.apps.myshop.data.db.remote.abstraction.notebook.RemoteEntryService
 import com.labs.devo.apps.myshop.data.models.notebook.Entry
@@ -12,16 +15,18 @@ import javax.inject.Inject
 
 class EntryRepositoryImpl
 @Inject constructor(
-    val localEntryService: LocalEntryService,
-    val remoteEntryService: RemoteEntryService
+    private val localEntryService: LocalEntryService,
+    private val remoteEntryService: RemoteEntryService
 ) : EntryRepository {
 
     override suspend fun getEntries(
         pageId: String,
         queryParams: QueryParams
     ): Flow<DataState<List<Entry>>> = flow {
+
         emit(DataState.loading<List<Entry>>(true))
         try {
+            checkPermissions(Permissions.GET_ENTRY)
             var localEntries = localEntryService.getEntries(pageId, queryParams)
             if (queryParams.whereQuery.isNotEmpty() && localEntries.isNullOrEmpty()) {
                 throw Exception("No record for this search query.")
@@ -42,6 +47,7 @@ class EntryRepositoryImpl
 
     override suspend fun insertEntries(entries: List<Entry>): DataState<List<Entry>> {
         return try {
+            checkPermissions(Permissions.CREATE_ENTRY)
             val insertedEntries = remoteEntryService.insertEntries(entries)
             val localInsertedEntries = localEntryService.insertEntries(insertedEntries)
             DataState.data(localInsertedEntries)
@@ -54,6 +60,7 @@ class EntryRepositoryImpl
 
     override suspend fun insertEntry(entry: Entry): DataState<Entry> {
         return try {
+            checkPermissions(Permissions.CREATE_ENTRY)
             val insertEntry = remoteEntryService.insertEntry(entry)
             val localInsertedEntry = localEntryService.insertEntry(insertEntry)
             DataState.data(localInsertedEntry)
@@ -66,6 +73,7 @@ class EntryRepositoryImpl
 
     override suspend fun updateEntries(entries: List<Entry>): DataState<List<Entry>> {
         return try {
+            checkPermissions(Permissions.CREATE_ENTRY)
             val updatedEntries = remoteEntryService.updateEntries(entries)
             val localUpdated = localEntryService.updateEntries(updatedEntries)
             DataState.data(localUpdated)
@@ -78,6 +86,7 @@ class EntryRepositoryImpl
 
     override suspend fun updateEntry(entry: Entry): DataState<Entry> {
         return try {
+            checkPermissions(Permissions.CREATE_ENTRY)
             val updatedEntry = remoteEntryService.updateEntry(entry)
             val localUpdatedEntry = localEntryService.updateEntry(updatedEntry)
             DataState.data(localUpdatedEntry)
@@ -90,6 +99,7 @@ class EntryRepositoryImpl
 
     override suspend fun deleteEntry(entry: Entry): DataState<Entry> {
         return try {
+            checkPermissions(Permissions.DELETE_ENTRY)
             remoteEntryService.deleteEntry(entry)
             localEntryService.deleteEntry(entry)
             DataState.data(entry)
@@ -102,6 +112,7 @@ class EntryRepositoryImpl
 
     override suspend fun deleteEntries(entries: List<Entry>): DataState<List<Entry>> {
         return try {
+            checkPermissions(Permissions.DELETE_ENTRY)
             remoteEntryService.deleteEntries(entries)
             localEntryService.deleteEntries(entries)
             DataState.data(entries)
